+++ CHANGELOG.md
@@ -6,7 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-[unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.1...HEAD
+[unreleased]: https://github.com/jannismain/mkdocs-macros-includex/compare/v0.0.2...HEAD
+
+## [0.0.2] - 2023-08-29
+[0.0.2]: https://github.com/jannismain/mkdocs-macros-includex/releases/tag/v0.0.2
+
+### Added
+
+- **caption**: add default or custom caption to included content in code blocks
+- **suffix**: add content after included content (but before any markup)
+
+### Changed
+
+- `includex` errors are raised by default
+- treat as error when no content is found to include
 
 ## [0.0.1] - 2023-07-21
 [0.0.1]: https://github.com/jannismain/mkdocs-macros-includex/releases/tag/v0.0.1